requirements.txt
streamlit==1.29.0,
pandas==2.2.0,
pydantic==2.6.0,
albumentations==1.3.1,
ipywidgets==8.1.1,
opencv_contrib_python==4.8.1.78,
opencv_python_headless==4.8.1.78,
tqdm==4.66.1,
torch==2.1.2,
Pillow==10.1.0,
numpy==1.26.3,
matplotlib==3.8.1,
Pillow==10.2.0,
torchvision==0.16.2,
uvicorn==0.27.0.post1,
ipython==8.16.1,
opencv_python==4.8.1.78,
fastapi==0.109.0,
auto_mix_prep==0.2.'